Add: response-helper 추가, 회원가입 API 적용 #11 Rename: prismaClient -> prisma-client 변경 #1
