+++ src/auth/auth.router.js
@@ -1,7 +1,7 @@
 import express from "express";
 import { AuthController } from "./auth.controller.js";
 import { AuthService } from "./auth.service.js";
-import { prisma } from "../configs/prismaClient.js";
+import { prisma } from "../configs/prisma-client.js";
 
 const authRouter = express.Router();
 const authService = new AuthService(prisma);
