+++ src/resumes/resume.router.js
@@ -1,7 +1,7 @@
 import express from "express";
 import { ResumeController } from "./resume.controller.js";
 import { ResumeService } from "./resume.service.js";
-import { prisma } from "../configs/prismaClient.js";
+import { prisma } from "../configs/prisma-client.js";
 
 const resumeService = new ResumeService(prisma);
 const resumeController = new ResumeController(resumeService);
